chore: logout all self sessions
