+++ apps/rest_api/lib/controllers/public/auth.ex
@@ -96,26 +96,4 @@ defmodule RestAPI.Controllers.Public.Auth do
       end
     end
   end
-
-  @doc "Logout the authenticated subject session."
-  @spec logout(conn :: Plug.Conn.t(), params :: map()) :: Plug.Conn.t()
-  def logout(%{private: %{session: session}} = conn, _params) do
-    session.jti
-    |> Authenticator.sign_out_session()
-    |> case do
-      {:ok, _any} -> send_resp(conn, :no_content, "")
-      {:error, _reason} = error -> error
-    end
-  end
-
-  @doc "Logout subject authenticated sessions."
-  @spec logout_all_sessions(conn :: Plug.Conn.t(), params :: map()) :: Plug.Conn.t()
-  def logout_all_sessions(%{private: %{session: session}} = conn, _params) do
-    session.subject_id
-    |> Authenticator.sign_out_all_sessions(session.subject_type)
-    |> case do
-      {:ok, _any} -> send_resp(conn, :no_content, "")
-      {:error, _reason} = error -> error
-    end
-  end
 end
