+++ apps/rest_api/lib/routers/public.ex
@@ -31,13 +31,6 @@ defmodule RestAPI.Routers.Public do
 
         post "/", Auth, :authorize
       end
-
-      scope "/" do
-        pipe_through :authenticated
-
-        post "/logout", Auth, :logout
-        post "/logout-all-sessions", Auth, :logout_all_sessions
-      end
     end
   end
 end
