+++ apps/rest_api/test/controllers/public/auth_test.exs
@@ -9,8 +9,6 @@ defmodule RestAPI.Controllers.Public.AuthTest do
   @content_type "application/x-www-form-urlencoded"
   @token_endpoint "/api/v1/auth/protocol/openid-connect/token"
   @authorization_code_endpoint "/api/v1/auth/protocol/openid-connect/authorize"
-  @logout_endpoint "/api/v1/auth/protocol/openid-connect/logout"
-  @logout_all_endpoint "/api/v1/auth/protocol/openid-connect/logout-all-sessions"
 
   describe "POST #{@authorization_code_endpoint}" do
     setup do
@@ -346,158 +344,6 @@ defmodule RestAPI.Controllers.Public.AuthTest do
                |> put_req_header("content-type", @content_type)
                |> post(@token_endpoint, %{"grant_type" => "authorization_code"})
                |> json_response(400)
-    end
-  end
-
-  describe "POST #{@logout_endpoint}" do
-    setup do
-      {:ok, access_token: "my-access-token", claims: default_claims()}
-    end
-
-    test "suceeds in if authenticated", %{conn: conn, access_token: access_token, claims: claims} do
-      expect(AuthenticatorMock, :validate_access_token, fn token ->
-        assert access_token == token
-        {:ok, claims}
-      end)
-
-      expect(AuthenticatorMock, :get_session, fn %{"jti" => jti} ->
-        assert claims["jti"] == jti
-        {:ok, success_session(claims)}
-      end)
-
-      expect(AuthenticatorMock, :sign_out_session, fn jti ->
-        assert claims["jti"] == jti
-        {:ok, %{}}
-      end)
-
-      assert conn
-             |> put_req_header("authorization", "Bearer #{access_token}")
-             |> post(@logout_endpoint)
-             |> response(204)
-    end
-
-    test "fails if sessions not active", %{conn: conn, access_token: access_token, claims: claims} do
-      expect(AuthenticatorMock, :validate_access_token, fn token ->
-        assert access_token == token
-        {:ok, claims}
-      end)
-
-      expect(AuthenticatorMock, :get_session, fn %{"jti" => jti} ->
-        assert claims["jti"] == jti
-        {:ok, success_session(claims)}
-      end)
-
-      expect(AuthenticatorMock, :sign_out_session, fn jti ->
-        assert claims["jti"] == jti
-        {:error, :not_active}
-      end)
-
-      assert conn
-             |> put_req_header("authorization", "Bearer #{access_token}")
-             |> post(@logout_endpoint)
-             |> response(403)
-    end
-
-    test "fails if session not found", %{conn: conn, access_token: access_token, claims: claims} do
-      expect(AuthenticatorMock, :validate_access_token, fn token ->
-        assert access_token == token
-        {:ok, claims}
-      end)
-
-      expect(AuthenticatorMock, :get_session, fn %{"jti" => jti} ->
-        assert claims["jti"] == jti
-        {:ok, success_session(claims)}
-      end)
-
-      expect(AuthenticatorMock, :sign_out_session, fn jti ->
-        assert claims["jti"] == jti
-        {:error, :not_found}
-      end)
-
-      assert conn
-             |> put_req_header("authorization", "Bearer #{access_token}")
-             |> post(@logout_endpoint)
-             |> response(404)
-    end
-  end
-
-  describe "POST #{@logout_all_endpoint}" do
-    test "suceeds in if authenticated", %{conn: conn} do
-      access_token = "my-access-token"
-      claims = default_claims()
-
-      expect(AuthenticatorMock, :validate_access_token, fn token ->
-        assert access_token == token
-        {:ok, claims}
-      end)
-
-      expect(AuthenticatorMock, :get_session, fn %{"jti" => jti} ->
-        assert claims["jti"] == jti
-        {:ok, success_session(claims)}
-      end)
-
-      expect(AuthenticatorMock, :sign_out_all_sessions, fn subject_id, subject_type ->
-        assert claims["sub"] == subject_id
-        assert claims["identity"] == subject_type
-        {:ok, 3}
-      end)
-
-      assert conn
-             |> put_req_header("authorization", "Bearer #{access_token}")
-             |> post(@logout_all_endpoint)
-             |> response(204)
-    end
-
-    test "fails if sesions not active", %{conn: conn} do
-      access_token = "my-access-token"
-      claims = default_claims()
-
-      expect(AuthenticatorMock, :validate_access_token, fn token ->
-        assert access_token == token
-        {:ok, claims}
-      end)
-
-      expect(AuthenticatorMock, :get_session, fn %{"jti" => jti} ->
-        assert claims["jti"] == jti
-        {:ok, success_session(claims)}
-      end)
-
-      expect(AuthenticatorMock, :sign_out_all_sessions, fn subject_id, subject_type ->
-        assert claims["sub"] == subject_id
-        assert claims["identity"] == subject_type
-        {:error, :not_active}
-      end)
-
-      assert conn
-             |> put_req_header("authorization", "Bearer #{access_token}")
-             |> post(@logout_all_endpoint)
-             |> response(403)
-    end
-
-    test "fails if session not found", %{conn: conn} do
-      access_token = "my-access-token"
-      claims = default_claims()
-
-      expect(AuthenticatorMock, :validate_access_token, fn token ->
-        assert access_token == token
-        {:ok, claims}
-      end)
-
-      expect(AuthenticatorMock, :get_session, fn %{"jti" => jti} ->
-        assert claims["jti"] == jti
-        {:ok, success_session(claims)}
-      end)
-
-      expect(AuthenticatorMock, :sign_out_all_sessions, fn subject_id, subject_type ->
-        assert claims["sub"] == subject_id
-        assert claims["identity"] == subject_type
-        {:error, :not_found}
-      end)
-
-      assert conn
-             |> put_req_header("authorization", "Bearer #{access_token}")
-             |> post(@logout_all_endpoint)
-             |> response(404)
     end
   end
